Add buttons and selection
